move fake searcher
